Fill `source` column with pattern from `source_id` only if necessary
Previously, the `source` column was always filled with a pattern from
the `source_id` column. Now, if the `source` column already exists and
contains a non-empty string, that value will be used instead. Therefore,
the pattern extraction now only serves as a backup. This allows us to
specify a `source` for each study in the meta data table without running
into an issue for studies in the public data (for which the `source` is
not defined in the study meta data).