Store `manhattan` in Parquet files partitioned by study
With our test studies, the `beta`/`beta_ci_lower`/`beta_ci_upper`
columns are missing so I added a line that checks whether those (or the
corresponding odds ratio columns) are missing and will initialize them
with null values if they are. When working with mixed studies (case and
continuous) this should not be an issue.